Update Group_Selected_Layers_in_Each_Artboard.js
Group was defaulting to 100x100, which was problematic while trying to batch update a bunch of shape layers in an icon library that all lived in a 32x32 artboard. I wanted the group to size to fit the contents of the group, not just be 100x100. I believe this fixes it. I was successfully able to update about 600 icons to use tints on the group containing the shape layer.